dpolya1 {FOO}	R Documentation

~~function to do ... ~~

Description

~~ A concise (1-5 lines) description of what the function does. ~~

Usage

dpolya1(partition, null, tau)

Arguments

partition 	 ~~Describe partition here~~

null 	 ~~Describe null here~~

tau 	 ~~Describe tau here~~

Details

~~ If necessary, more details than the __description__ above ~~

Value

~Describe the value returned If it is a LIST, use

comp1 	 Description of `comp1'

comp2 	 Description of `comp2'

...

WARNING

....

Note

~~further notes~~

Author(s)

~~who you are~~

References

~put references to the literature/web site here ~

See Also

~~objects to SEE ALSO as ~~fun~~, ~~~

Examples

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--         or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function(partition,null,tau)
 {
  #####################################################################
  #
  # Evaluate the log prior mass of a Double Polya Prior-augmented state.
  #
  # This is the prior for an augmented-state partition in ISN model
  # (Used in the data-augmentation part of the MCMC algorithm.)
  #
  #
  # partition = nswitch vector holding distinct path labels
  # null = logical vector indicating active (i.e. relevant) switches
  # tau = clustering parameter: bigger <-> more paths
  #
  #####################################################################

  nswitch <- length(partition)

  # The partition componenent
  incidence <- outer( partition, unique(partition), "==" )
  sizes <- t(incidence) 
  npath <- length(sizes)
  logp <-  npath*log(tau) + sum( lgamma(sizes) ) 
  logp <- logp + lgamma(tau)-lgamma(nswitch+tau)

  # The activation component
  nnull <- sum( null )
  logp <- logp+lgamma(nnull+1)+lgamma(nswitch-nnull+1)-lgamma(nswitch+2)
  return( logp )
 }